Update 19 Maximum product subarray.java

+++ Arrays/19 Maximum product subarray.java
@@ -10,9 +10,9 @@ Sample Output 1:
 Solution: 
 public class Solution {
     public static int subarrayWithMaxProduct(int []arr){
-        int pre = 1; int suf =1;     //initialize pre and suff as 1;
-        int ans = Integer.MIN_VALUE; 
-        int n = arr.length;
+        double pre = 1; int suf =1;     //initialize pre and suff as 1;
+        double ans = Integer.MIN_VALUE; 
+        double n = arr.length;
         for(int i =0; i<n;i++){
             if(pre==0) pre = 1;     //if pre and suff becomes 0 then reset the value to 1; 
             if(suf ==0) suf =1;
@@ -20,6 +20,6 @@ public class Solution {
             suf = suf*arr[n-i-1];   //suff multiplies from end of the array
             ans = Math.max(ans,Math.max(pre,suf)); //compare the pre and suff max, then compare it with ans.
          }
-            return ans;    //return the max value
+            return (int)ans;    //return the max value
     }
 }
